Add shebang, env, formatting, minor fixes
With the shebang the script can directly be called in bash.
dotenv allows the loading of variables from .env files,
this comes in handy when working with global instances, like
a database name that might change depending on environment.
With Rubocop comes some ruling to the formatting of Ruby
code. Fix paths.