Create deployment action
This will upload:

* docker-compose.yml
* prometheus.yml

to the production server and place it into
a folder there. These files will then be used
after Travis has built and download the latest
app and API images.
This takes care of the problem of not being able
to docker-compose with our current deployment
of Docker Hub and Travis.